Change normalization behavior of `showImage`
The `showImage` method used to call `vals2Grayscale` every time an image
was to be shown. This is incorrect, as the original image after being
loaded should be unaltered. Now `showImage` has a `normalize` parameter,
which defaults to True (this is the most frequent use-case), but is
called with False when the original image should be shown.