Rework the broken image transforms and how the image is shown
Now the `self.currentImage` is updated in `showImage`, which means that
everytime an image is shown, the current image is saved without having
to add a separate command after a call of `showImage`. By using
`self.currentImage` in the transformations, the image skewing has been
fixed. The skewing was happening probably due to incorrect QImage to
numpy array conversion. Now this conversion has been eliminated by using
`self.currentImage`.